DataSeeding

+++ src/Services/Catalog/Catalog.API/Products/DeleteProduct/DeleteProductHandler.cs
@@ -9,7 +9,11 @@ namespace Catalog.API.Products.DeleteProduct
     {
         public async Task<DeleteProductResult> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
         {
-           
+            var product = await session.LoadAsync<Product>(command.Id, cancellationToken);
+            if (product is null)
+            {
+                throw new ProductNotFoundException(command.Id);
+            }
             session.Delete<Product>(command.Id);
             await session.SaveChangesAsync(cancellationToken);
             return new DeleteProductResult(true);
